// daisy_scout_web/db/models/pickList-2025-data.tsx
export type PickList2025Data = {
  eventId: string;
  teamNumber: number;
  pickIndex: number;
  allianceSelected: boolean;
  rank: number;
  autoStartLocAlliance: number;
  autoStartLocCenter: number;
  autoStartLocOpposing: number;
  autoStartLocNoShow: number;
  autoStart: string;
  autoReefL1: number;
  autoReefL2: number;
  autoReefL3: number;
  autoReefL4: number;
  autoCoral: number;
  autoLevel: string;
  coralA: number;
  coralB: number;
  coralC: number;
  coralD: number;
  coralE: number;
  coralF: number;
  autoReef: string;
  leave: number;
  autoNet: number;
  autoProcessor: number;
  teleopReefL1: number;
  teleopReefL2: number;
  teleopReefL3: number;
  teleopReefL4: number;
  teleopCoral: number;
  teleopLevel: string;
  teleopProcessor: number;
  teleopNet: number;
  teleopAlgaeRemoved: number;
  endgamePark: number;
  endgameShallow: number;
  endgameDeep: number;
  endgameNone: number;
  endgame: string;
  driverAbility: number | null | undefined;
  operability: string | null | undefined;
  role: string | null | undefined;
  hpNetShots: string | null | undefined;
  hpNetMade: number | null | undefined;
  driveTrain: string | null | undefined;
};